Moved service request form to root modal dialog
