+++ layout/AppMenu.js
@@ -1,12 +1,15 @@
 import getConfig from 'next/config';
-import React, { useContext } from 'react';
+import React, { useContext, useState } from 'react';
 import AppMenuitem from './AppMenuitem';
 import { LayoutContext } from './context/layoutcontext';
 import { MenuProvider } from './context/menucontext';
+import ServiceRequestDialog from '@components/serviceRequest/ServiceRequestDialog';
 // import pawIcon from '/images/paw.svg';
 
 const AppMenu = () => {
     const { layoutConfig } = useContext(LayoutContext);
+    const [dialog, showDialog] = useState('')
+    const onClose = () => showDialog('')
     const contextPath = getConfig().publicRuntimeConfig.contextPath;
     const model = [
         {
@@ -16,12 +19,19 @@ const AppMenu = () => {
                 { label: 'Clients', icon: 'pi pi-fw pi-users', to: '/clients' },
                 { label: 'Animals', iconSrc: `${contextPath}/images/paw.svg`, to: '/animals' },
                 { label: 'Reports', icon: 'pi pi-fw pi-book', to: '/reports' },
-                { label: 'New Request', iconSrc: `${contextPath}/images/pawClipboard.svg`, to: '/serviceRequest/create' }
+                { separator: true },
+                { 
+                  label: 'New Request', 
+                  iconSrc: `${contextPath}/images/pawClipboard.svg`, 
+                  key: 'newServiceRequest', 
+                  command: () => {if (!dialog) showDialog('newServiceRequest')} 
+                }
             ]
         }
     ];
 
     return (
+      <>
         <MenuProvider>
             <ul className="layout-menu">
                 {model.map((item, i) => {
@@ -29,6 +39,11 @@ const AppMenu = () => {
                 })}
             </ul>
         </MenuProvider>
+        <ServiceRequestDialog 
+          visible={dialog === 'newServiceRequest'}
+          onClose={onClose}
+        />
+      </>
     );
 };
