--- pages/serviceRequest/create/index.tsx
@@ -1,7 +1,0 @@
-import ServiceRequestForm from "../../../src/components/serviceRequest/ServiceRequestForm";
-
-function ServiceRequest() {
-  return <ServiceRequestForm />
-}
-
-export default ServiceRequest
